+++ index.html
@@ -250,7 +250,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="./js/main.js" async></script>
+  <script src="./js/main.js" type="module" async></script>
 </body>
 
 </html>
