+++ js/const.js
@@ -1,0 +1,99 @@
+const DEFAULT_PRECISION = 0;
+
+const AVATARS = [
+  'img/avatars/user01.png',
+  'img/avatars/user02.png',
+  'img/avatars/user03.png',
+  'img/avatars/user04.png',
+  'img/avatars/user05.png',
+  'img/avatars/user06.png',
+  'img/avatars/user07.png',
+  'img/avatars/user08.png',
+  'img/avatars/user09.png',
+  'img/avatars/user10.png',
+];
+
+const TITLES = [
+  'Заголовок 1',
+  'Заголовок 2',
+  'Заголовок 3',
+  'Заголовок 4',
+  'Заголовок 5',
+];
+
+const ADRESS = [
+  'location.lat',
+  'location.lng',
+];
+
+const PRICES = [
+  2000,
+  3000,
+  4000,
+  5000,
+];
+
+const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+
+const ROOMS = [
+  1,
+  2,
+  3,
+];
+
+const GUESTS = [
+  1,
+  2,
+  3,
+  4,
+];
+
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+
+const DESCRIPTION = [
+  'Хорошая,чистая',
+  'Средней комфортности',
+  'Лучше не заезжать',
+  'Берите, не пожалеете',
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+
+const LAT = [
+  35.65000,
+  35.70000,
+];
+
+const LNG = [
+  139.70000,
+  139.80000,
+];
